.agent/skills/note-thumbnail-generator/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
requests
python-dotenv
google-generativeai
pyngrok

